+++ tools/machine-learning/mujoco/packages/mujoco-interactive-viewer/src/mujoco_interactive_viewer/marker.py
@@ -1,0 +1,14 @@
+from dataclasses import dataclass
+
+import mujoco
+import numpy as np
+from numpy.typing import NDArray
+
+
+@dataclass
+class Marker:
+    kind: mujoco.mjtGeom
+    size: NDArray[np.float64]
+    position: NDArray[np.float64]
+    material: NDArray[np.float64]
+    rgba: NDArray[np.float32]
