+++ tools/machine-learning/mujoco/packages/mujoco-interactive-viewer/src/mujoco_interactive_viewer/overlay.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Overlay:
+    text1: str
+    text2: str
+
+    def __init__(self) -> None:
+        self.text1 = ""
+        self.text2 = ""
